extracted the CommitPair class as a public class
In later commits it will be used by other classes
